pyproject.toml
[tool.poetry]
name = "osu-receptor"
version = "2.0.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
pillow = "^11.1.0"
pyinstaller = "^6.12.0"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
